requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altgraph=0.17.3=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py310h4e76f89_0
brotlipy=0.7.0=py310hca72f7f_1002
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.01.10=hecd8cb5_0
certifi=2022.12.7=py310hecd8cb5_0
cffi=1.15.1=py310h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310hecd8cb5_0
cryptography=38.0.4=py310hf6deb26_0
ffmpeg=4.3=h0a44026_0
flask=2.2.2=py310hecd8cb5_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=3.4=py310hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py310hecd8cb5_0
jpeg=9e=hca72f7f_0
keras=2.11.0=pyhd8ed1ab_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libffi=3.4.2=hecd8cb5_6
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.5.0=h2cd0358_0
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
macholib=1.16.2=pypi_0
markupsafe=2.1.1=py310hca72f7f_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
ncurses=6.3=hca72f7f_3
nettle=3.7.3=h230ac6f_1
numexpr=2.8.4=py310h9638375_0
numpy=1.23.5=py310h9638375_0
numpy-base=1.23.5=py310ha98c3c9_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1t=hca72f7f_0
packaging=22.0=py310hecd8cb5_0
pandas=1.5.2=py310h3ea8b11_0
pillow=9.3.0=py310h81888ad_1
pip=22.3.1=py310hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyinstaller=5.8.0=pypi_0
pyinstaller-hooks-contrib=2023.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyside6=6.4.2=pypi_0
pyside6-addons=6.4.2=pypi_0
pyside6-essentials=6.4.2=pypi_0
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.8=h218abb5_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.13.1=py3.10_0
pytz=2022.7=py310hecd8cb5_0
readline=8.2=hca72f7f_0
requests=2.28.1=py310hecd8cb5_0
setuptools=65.6.3=py310hecd8cb5_0
shiboken6=6.4.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h880c91c_0
tk=8.6.12=h5d9f67b_0
torchaudio=0.13.1=py310_cpu
torchvision=0.14.1=py310_cpu
typing_extensions=4.4.0=py310hecd8cb5_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=py310hecd8cb5_0
werkzeug=2.2.2=py310hecd8cb5_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
